# internal/api/router.go
package api

import (
	"net/http"
	"os"

	"github.com/go-chi/chi/v5"
	chiMiddleware "github.com/go-chi/chi/v5/middleware"
	"github.com/prometheus/client_golang/prometheus/promhttp"
	"github.com/tahsinarafat/aioj/internal/api/handler"
	"github.com/tahsinarafat/aioj/internal/api/middleware"
	"github.com/tahsinarafat/aioj/internal/auth"
)

func NewRouter(d Deps, jwtManager *auth.JWTManager) http.Handler {
	authH := d.Auth
	problemH := d.Problem
	problemI18nH := d.ProblemI18n
	submissionH := d.Submission
	contestH := d.Contest
	contestProblemH := d.ContestProblem
	vjudgeH := d.VJudge
	adminH := d.Admin
	testcaseH := d.Testcase
	wsManager := d.WS
	ratingH := d.Rating
	registrationH := d.Registration
	virtualH := d.Virtual
	gymH := d.Gym
	hackH := d.Hack
	statsH := d.Stats
	notifH := d.Notification
	groupH := d.Group
	teamH := d.Team
	blogH := d.Blog
	editorialH := d.Editorial
	apiKeyH := d.APIKey
	webhookH := d.Webhook
	recommendationH := d.Recommendation
	rankingsH := d.Rankings
	usersH := d.Users
	searchH := d.Search
	langLimitH := d.LangLimit
	importH := d.Import
	orgH := d.Org
	classH := d.Class
	trainingH := d.Training
	plagiarismH := d.Plagiarism
	mediaH := d.Media
	onsiteH := d.Onsite
	onsiteBatchH := d.OnsiteBatch
	clarificationH := d.Clarification
	noticeH := d.Notice
	botAccountH := d.BotAccount
	settingsH := d.Settings
	langAdminH := d.LangAdmin
	remoteLangH := d.RemoteLang
	adminSubH := d.AdminSub
	backupH := d.Backup
	r := chi.NewRouter()
	rl := middleware.NewRateLimiter()
	r.Use(
		middleware.SecurityHeaders(),
		middleware.CSRF(d.CSRFSecret),
		middleware.RateLimit(rl, "/api/health", "/api/ws", "/metrics"),
		chiMiddleware.RequestID,
		chiMiddleware.RealIP,
		middleware.Logging,
		middleware.SentryRecover,
		chiMiddleware.Recoverer,
	)

	r.Get("/metrics", promhttp.Handler().ServeHTTP)

	if d.Health != nil {
		r.Get("/api/health", d.Health.Health)
	} else {
		r.Get("/api/health", func(w http.ResponseWriter, _ *http.Request) { w.Write([]byte(`{"status":"ok"}`)) })
	}

	if d.Activity != nil {
		r.Get("/api/activity", d.Activity.List)
	}

	// Public status page (Phase C/D) — static HTML polling /api/health.
	r.Get("/status", func(w http.ResponseWriter, _ *http.Request) {
		b, err := os.ReadFile("deploy/status.html")
		if err != nil {
			http.Error(w, "status page missing", http.StatusNotFound)
			return
		}
		w.Header().Set("Content-Type", "text/html; charset=utf-8")
		w.Header().Set("Cache-Control", "public, max-age=30")
		_, _ = w.Write(b)
	})

	if d.Legal != nil {
		r.Get("/api/legal/{doc}", d.Legal.Serve)
	}

	// Public sitemap for crawlers. Built from the same public-visibility rules
	// the problem/contest APIs apply, so it never advertises an auth-gated URL.
	if d.Sitemap != nil {
		r.Get("/sitemap.xml", func(w http.ResponseWriter, req *http.Request) {
			handler.ServeSitemapSections(d.Sitemap(req.Context()), w, req)
		})
	}

	if d.ContestFeed != nil {
		r.Get("/feed/contests.atom", func(w http.ResponseWriter, req *http.Request) {
			items, err := d.ContestFeed(req.Context())
			handler.ServeAtomFeed(items, err,
				"AIOJ — Contests", "Recent and upcoming contests on AIOJ",
				"http://localhost:8081/feed/contests.atom",
				"http://localhost:8081/feed/contests.atom", w, req)
		})
	}

	if d.CommentFeed != nil {
		r.Get("/feed/comments.atom", func(w http.ResponseWriter, req *http.Request) {
			items, err := d.CommentFeed(req.Context())
			handler.ServeAtomFeed(items, err,
				"AIOJ — Comments", "The latest comments on AIOJ",
				"http://localhost:8081/feed/comments.atom",
				"http://localhost:8081/feed/comments.atom", w, req)
		})
	}

	if d.Feed != nil {
		r.Get("/feed/problems.atom", func(w http.ResponseWriter, req *http.Request) {
			items, err := d.Feed(req.Context())
			handler.ServeAtomFeed(items, err,
				"AIOJ — New Problems", "The latest problems added to AIOJ",
				"http://localhost:8081/feed/problems.atom",
				"http://localhost:8081/feed/problems.atom", w, req)
		})
	}

	if d.BlogFeed != nil {
		r.Get("/feed/blog.atom", func(w http.ResponseWriter, req *http.Request) {
			items, err := d.BlogFeed(req.Context())
			handler.ServeAtomFeed(items, err,
				"AIOJ — Blog", "The latest posts from AIOJ",
				"http://localhost:8081/feed/blog.atom",
				"http://localhost:8081/feed/blog.atom", w, req)
		})
	}

	r.Get("/api/search", searchH.Search)

	r.With(rl.StrictAuth).Post("/api/auth/register", authH.Register)
	r.With(rl.StrictAuth).Post("/api/auth/login", authH.Login)
	r.Post("/api/auth/refresh", authH.Refresh)
	r.With(rl.StrictAuth).Post("/api/auth/forgot-password", authH.ForgotPassword)
	r.With(rl.StrictAuth).Post("/api/auth/reset-password", authH.ResetPassword)
	if d.VerifyEmail != nil {
		r.Get("/api/auth/verify-email/{token}", d.VerifyEmail.Verify)
	}
	r.With(middleware.AuthMiddleware(jwtManager)).Post("/api/auth/verify-email/resend", authH.ResendVerification)
	if d.TwoFA != nil {
		r.With(middleware.AuthMiddleware(jwtManager)).Post("/api/auth/2fa/begin", d.TwoFA.Begin)
		r.With(middleware.AuthMiddleware(jwtManager)).Post("/api/auth/2fa/enable", d.TwoFA.Enable)
		r.With(middleware.AuthMiddleware(jwtManager)).Post("/api/auth/2fa/disable", d.TwoFA.Disable)
	}
	if d.TwoFAVerify != nil {
		r.With(rl.StrictAuth).Post("/api/auth/2fa/verify", d.TwoFAVerify.Verify)
	}
	if d.DevMail != nil {
		r.Get("/api/dev/mail", d.DevMail.List)
		r.Delete("/api/dev/mail", d.DevMail.Clear)
	}
	if d.OAuthStart != nil {
		r.Get("/api/auth/oauth/{provider}/start", d.OAuthStart.Start)
	}
	if d.OAuthCallback != nil {
		r.Get("/api/auth/oauth/{provider}/callback", d.OAuthCallback.Callback)
	}

	r.Get("/api/users/{username}", usersH.GetByUsername)
	r.Get("/api/users/{username}/submissions", usersH.GetUserSubmissions)
	r.Get("/api/users/{username}/blogs", usersH.GetUserBlogs)
	r.Get("/api/users/{username}/comments", usersH.GetUserComments)

	// Phase D: achievements + social graph
	if d.Achievements != nil {
		r.With(middleware.AuthMiddleware(jwtManager)).Get("/api/users/{userId}/achievements", d.Achievements.List)
		r.With(middleware.AuthMiddleware(jwtManager)).Post("/api/users/{userId}/achievements/{code}", d.Achievements.Award)
	}
	if d.Friendships != nil {
		r.With(middleware.AuthMiddleware(jwtManager)).Post("/api/users/{username}/follow", d.Friendships.Follow)
		r.With(middleware.AuthMiddleware(jwtManager)).Delete("/api/users/{username}/follow", d.Friendships.Unfollow)
		r.Get("/api/users/{username}/follow/status", d.Friendships.Status)
		r.Get("/api/users/{username}/followers", d.Friendships.ListFollowers)
		r.Get("/api/users/{username}/following", d.Friendships.ListFollowing)
	}
	if d.UsersExport != nil {
		r.With(middleware.AuthMiddleware(jwtManager)).Get("/api/users/me/export", d.UsersExport.ExportMyData)
	}
	if d.UsersDeletion != nil {
		r.With(middleware.AuthMiddleware(jwtManager)).Delete("/api/users/me", d.UsersDeletion.DeleteAccount)
	}
	r.With(middleware.AuthMiddleware(jwtManager)).Get("/api/users/profile/edit", usersH.GetProfile)
	r.With(middleware.AuthMiddleware(jwtManager)).Put("/api/users/profile/edit", usersH.UpdateProfile)
	r.With(middleware.AuthMiddleware(jwtManager)).Put("/api/users/profile/password", usersH.UpdatePassword)
	r.With(middleware.AuthMiddleware(jwtManager)).Get("/api/users/profile/invites", usersH.GetUserPendingInvites)

	r.Route("/api/problems", func(r chi.Router) {
		r.Get("/", problemH.List)
		r.Get("/tags", problemH.ListTags)
		r.With(middleware.AuthMiddleware(jwtManager)).Get("/my", problemH.ListMyProblems)
		r.With(middleware.OptionalAuthMiddleware(jwtManager)).Get("/{slug}", problemH.GetBySlug)

		// Per-problem statement translations (dmoj-style i18n). Reads are public
		// so a solver can fetch the statement in their own language; writes sit
		// behind auth like the other setter operations.
		r.Route("/{id}/i18n", func(r chi.Router) {
			r.Get("/", problemI18nH.List)
			r.With(middleware.AuthMiddleware(jwtManager)).Put("/{lang}", problemI18nH.Upsert)
			r.With(middleware.AuthMiddleware(jwtManager)).Delete("/{lang}", problemI18nH.Delete)
		})
		r.Get("/{slug}/submissions", submissionH.ListByProblem)
		r.Get("/{slug}/language-limits", langLimitH.List)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", problemH.Create)
			r.Put("/{slug}", problemH.Update)
			r.Delete("/{slug}", problemH.Delete)
			r.Get("/{slug}/permissions", problemH.ListPermissions)
			r.Post("/{slug}/permissions", problemH.AddPermission)
			r.Delete("/{slug}/permissions/{userId}", problemH.RemovePermission)
			r.Post("/{slug}/testcases", testcaseH.Upload)
			r.Post("/{slug}/language-limits", langLimitH.Set)
			r.Delete("/{slug}/language-limits/{lang}", langLimitH.Delete)
			r.Get("/{slug}/export", problemH.Export)
			r.Post("/import", importH.Import)
			r.Post("/import/codeforces", importH.ImportCodeforces)
			r.Post("/import/cses", importH.ImportCSES)
			r.Post("/import/atcoder", importH.ImportAtCoder)
			r.Post("/import/toph", importH.ImportToph)
			r.Post("/import/qoj", importH.ImportQOJ)
			r.Post("/{slug}/media", mediaH.Upload)
		})
	})

	r.Route("/api/submissions", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Post("/", submissionH.Create)
		r.Post("/upsolving", submissionH.CreateUpsolving)
		r.Post("/run", submissionH.CustomRun)
		r.Get("/", submissionH.ListByUser)
		r.Get("/{id}", submissionH.GetByID)
		r.Post("/{id}/retry-remote", submissionH.RetryRemote)
		r.Post("/{id}/recheck-remote", submissionH.RecheckRemote)
	})

	r.Route("/api/contests", func(r chi.Router) {
		r.Get("/", contestH.List)
		r.Get("/formats", contestH.ListAvailableFormats)
		r.With(middleware.AuthMiddleware(jwtManager)).Get("/my", contestH.ListMyContests)
		r.With(middleware.OptionalAuthMiddleware(jwtManager)).Get("/{id}", contestH.GetByID)
		r.With(middleware.OptionalAuthMiddleware(jwtManager)).Get("/{id}/scoreboard", contestH.Scoreboard)
		r.With(middleware.OptionalAuthMiddleware(jwtManager)).Get("/{id}/stats", contestH.ContestStats)
		r.With(middleware.OptionalAuthMiddleware(jwtManager)).Get("/{contestId}/problems/{index}", contestProblemH.GetByIndex)
		r.With(middleware.OptionalAuthMiddleware(jwtManager)).Get("/{id}/pdf", contestH.DownloadPDF)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", contestH.Create)
			r.Post("/educational", contestH.CreateEducational)
			r.Put("/{id}", contestH.Update)
			r.Delete("/{id}", contestH.Delete)
			r.Get("/{id}/permissions", contestH.ListPermissions)
			r.Post("/{id}/permissions", contestH.AddPermission)
			r.Delete("/{id}/permissions/{userId}", contestH.RemovePermission)
			r.Post("/{id}/problems", contestH.AddProblem)
			r.Put("/{id}/problems/{problemId}", contestH.UpdateProblem)
			r.Delete("/{id}/problems/{problemId}", contestH.RemoveProblem)
			r.Get("/{id}/submissions", submissionH.ListByContest)
			r.Post("/{id}/calculate-ratings", contestH.CalculateRatings)
		})
		r.With(middleware.AuthMiddleware(jwtManager)).Post("/{id}/register-team", contestH.RegisterTeam)
		r.Get("/{id}/team-registrations", contestH.ListTeamRegistrations)
	})

	r.Route("/api/contests/{contestId}/clarifications", func(r chi.Router) {
		r.Get("/", clarificationH.List)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", clarificationH.Create)
			r.Post("/{id}/answer", clarificationH.Answer)
		})
	})

	r.Route("/api/contests/{contestId}/notices", func(r chi.Router) {
		r.Get("/", noticeH.List)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", noticeH.Create)
			r.Delete("/{id}", noticeH.Delete)
		})
	})

	r.Route("/api/contests/{id}/register", func(r chi.Router) {
		r.Get("/", registrationH.CheckRegistration)
		r.With(middleware.AuthMiddleware(jwtManager)).Post("/", registrationH.Register)
		r.With(middleware.AuthMiddleware(jwtManager)).Delete("/", registrationH.Unregister)
	})

	r.Get("/api/contests/{id}/registrations", registrationH.ListRegistrations)

	r.Route("/api/contests/{id}/onsite", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Get("/balloons", onsiteH.ListBalloons)
		r.Post("/balloons/{balloonId}/dispatch", onsiteH.DispatchBalloon)
		r.Post("/print", onsiteH.RequestPrint)
		r.Get("/prints", onsiteH.ListPrints)
		r.Post("/prints/{printId}/status", onsiteH.UpdatePrintStatus)
		r.Post("/generate", onsiteBatchH.GenerateBatch)
		r.Get("/users", onsiteBatchH.ListBatch)
		r.Delete("/users/{userId}", onsiteBatchH.DeleteUser)
	})

	r.Post("/api/contests/{id}/onsite/login", onsiteBatchH.LoginAsTeam)

	r.Route("/api/vjudge", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Get("/bots", vjudgeH.ListBots)
		r.Post("/submit", vjudgeH.Submit)
	})

	r.Route("/api/admin", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Use(middleware.RequireRole("admin"))
		if d.AuditLog != nil {
			am := &middleware.AuditRecorder{Store: d.AuditLog.Audit}
			r.Use(am.Middleware)
		}
		r.Get("/users", adminH.ListUsers)
		r.Put("/users/{id}/role", adminH.UpdateUserRole)
		r.Get("/setter-applications", adminH.ListSetterApps)
		r.Post("/setter-applications/{id}/review", adminH.ReviewSetterApp)

		r.Route("/bot-accounts", func(r chi.Router) {
			r.Get("/", botAccountH.List)
			r.Post("/", botAccountH.Create)
			r.Post("/test-login", botAccountH.TestLogin)
			r.Get("/{id}", botAccountH.GetByID)
			r.Put("/{id}", botAccountH.Update)
			r.Delete("/{id}", botAccountH.Delete)
		})

		r.Route("/settings", func(r chi.Router) {
			r.Get("/", settingsH.List)
			r.Post("/", settingsH.Set)
			r.Get("/{key}", settingsH.Get)
			r.Delete("/{key}", settingsH.Delete)
		})

		if d.CDN != nil {
			r.Get("/cdn/status", d.CDN.Status)
			r.Post("/cdn/purge", d.CDN.Purge)
		}

		if d.AuditLog != nil {
			r.Get("/audit-log", d.AuditLog.List)
		}

		r.Route("/languages", func(r chi.Router) {
			r.Get("/", langAdminH.List)
			r.Post("/", langAdminH.Create)
			r.Get("/detect", langAdminH.Detect)
			r.Get("/templates", langAdminH.Templates)
			r.Get("/{key}", langAdminH.Get)
			r.Put("/{key}", langAdminH.Update)
			r.Delete("/{key}", langAdminH.Delete)
			r.Get("/{key}/raw", langAdminH.GetRaw)
			r.Put("/{key}/raw", langAdminH.UpdateRaw)
			r.Post("/{key}/test", langAdminH.Test)
		})

		r.Route("/remote-languages", func(r chi.Router) {
			r.Get("/{platform}", remoteLangH.ListByPlatform)
			r.Post("/", remoteLangH.Create)
			r.Put("/{id}", remoteLangH.Update)
			r.Delete("/{id}", remoteLangH.Delete)
			r.Post("/detect/{platform}", remoteLangH.AutoDetect)
			r.Post("/bulk", remoteLangH.BulkUpsert)
		})

		r.Route("/submissions", func(r chi.Router) {
			r.Get("/pending-remote", adminSubH.ListPendingRemote)
			r.Post("/{id}/rejudge", adminSubH.Rejudge)
			r.Post("/{id}/refresh", adminSubH.ForceRefresh)
		})

		r.Route("/backups", func(r chi.Router) {
			r.Get("/", backupH.List)
			r.Post("/", backupH.Create)
			r.Post("/upload", backupH.Upload)
			r.Get("/{filename}", backupH.Download)
			r.Delete("/{filename}", backupH.Delete)
			r.Post("/{filename}/restore", backupH.Restore)
		})

		r.Route("/ai-models", func(r chi.Router) {
			r.Get("/", d.AIModel.List)
			r.Post("/", d.AIModel.Create)
			r.Post("/test", d.AIModel.TestConnection)
			r.Get("/{id}", d.AIModel.GetByID)
			r.Put("/{id}", d.AIModel.Update)
			r.Put("/{id}/toggle", d.AIModel.Toggle)
			r.Delete("/{id}", d.AIModel.Delete)
		})
	})

	r.Get("/api/remote-languages/{platform}", remoteLangH.ListByPlatform)

	r.Group(func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Post("/api/auth/setter-apply", adminH.ApplySetter)
		r.Get("/api/auth/setter-status", adminH.GetSetterStatus)
	})

	r.Get("/api/ws", wsManager.Handle)

	r.Route("/api/rating", func(r chi.Router) {
		r.Get("/user/{userId}", ratingH.GetByUser)
		r.Get("/contest/{contestId}", ratingH.GetByContest)
	})

	r.Get("/api/rankings", rankingsH.List)

	r.Route("/api/virtual", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Post("/start", virtualH.Start)
		r.Get("/status", virtualH.Status)
		r.Post("/{id}/complete", virtualH.Complete)
	})

	r.Route("/api/gym", func(r chi.Router) {
		r.Get("/", gymH.List)
		r.Get("/{id}", gymH.GetByID)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", gymH.Create)
			r.Post("/{id}/solve", gymH.MarkSolved)
			r.Get("/{id}/solved", gymH.IsSolved)
		})
	})

	r.Route("/api/hacks", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Post("/", hackH.SubmitHack)
		r.Get("/{id}", hackH.GetHack)
		r.Get("/contest/{contestId}", hackH.ListContestHacks)
		r.Get("/hackable/{contestId}/{problemId}", hackH.ListHackableSubmissions)
	})

	r.Route("/api/stats", func(r chi.Router) {
		r.Get("/platform", statsH.GetPlatformStats)
		r.Get("/problems/{problemId}", statsH.GetProblemStats)
		r.With(middleware.AuthMiddleware(jwtManager)).Get("/me", statsH.GetUserStats)
	})

	r.Route("/api/notifications", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Get("/", notifH.List)
		r.Get("/unread-count", notifH.UnreadCount)
		r.Post("/{id}/read", notifH.MarkAsRead)
		r.Post("/read-all", notifH.MarkAllAsRead)
		r.Get("/preferences", notifH.GetPreferences)
		r.Put("/preferences", notifH.UpdatePreferences)
	})

	r.Route("/api/groups", func(r chi.Router) {
		r.Get("/", groupH.List)
		r.With(middleware.AuthMiddleware(jwtManager)).Post("/join-code", groupH.JoinByCode)
		r.Get("/{id}", groupH.GetByID)
		r.Get("/{id}/members", groupH.GetMembers)
		r.Get("/{id}/contests", groupH.GetContests)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", groupH.Create)
			r.Get("/my", groupH.ListByUser)
			r.Put("/{id}", groupH.Update)
			r.Delete("/{id}", groupH.Delete)
			r.Post("/{id}/join", groupH.Join)
			r.Post("/{id}/leave", groupH.Leave)
			r.Post("/{id}/contests", groupH.AddContest)
			r.Delete("/{id}/contests/{contestId}", groupH.RemoveContest)
			r.Post("/{id}/invite", groupH.InviteMember)
			r.Post("/{id}/respond", groupH.RespondInviteRequest)
			r.Get("/{id}/pending", groupH.GetPendingMembers)
		})
	})

	r.Route("/api/teams", func(r chi.Router) {
		r.Get("/", teamH.List)
		r.Get("/{id}", teamH.GetByID)
		r.Get("/{id}/members", teamH.GetMembers)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", teamH.Create)
			r.Post("/{id}/join", teamH.Join)
			r.Post("/{id}/leave", teamH.Leave)
			r.Put("/{id}", teamH.UpdateTeam)
			r.Delete("/{id}", teamH.DeleteTeam)
			r.Post("/{id}/invite", teamH.InviteMember)
			r.Post("/{id}/request", teamH.RequestJoin)
			r.Post("/{id}/respond", teamH.RespondInviteRequest)
			r.Get("/{id}/pending", teamH.GetPendingMembers)
		})
	})

	r.Route("/api/blog", func(r chi.Router) {
		r.Get("/", blogH.ListPosts)
		r.Get("/{id}", blogH.GetPost)
		r.Get("/{type}/{id}/comments", blogH.GetComments)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", blogH.CreatePost)
			r.Put("/{id}", blogH.UpdatePost)
			r.Delete("/{id}", blogH.DeletePost)
			r.Get("/{id}/vote", blogH.GetUserVote)
			r.Post("/comments", blogH.CreateComment)
			r.Post("/vote", blogH.Vote)
		})
	})

	r.Route("/api/editorials", func(r chi.Router) {
		r.Get("/", editorialH.List)
		r.Get("/{id}", editorialH.GetByID)
		r.Get("/problem/{problemId}", editorialH.GetByProblem)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", editorialH.Create)
		})
	})

	r.Route("/api/keys", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Get("/", apiKeyH.List)
		r.Post("/", apiKeyH.Create)
		r.Delete("/{id}", apiKeyH.Delete)
	})

	r.Route("/api/webhooks", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Get("/", webhookH.List)
		r.Post("/", webhookH.Create)
		r.Delete("/{id}", webhookH.Delete)
	})

	r.Route("/api/recommendations", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Get("/", recommendationH.GetRecommendations)
	})

	r.Route("/api/organizations", func(r chi.Router) {
		r.Get("/", orgH.List)
		r.Get("/{id}", orgH.GetByID)
		r.Get("/{id}/members", orgH.GetMembers)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", orgH.Create)
			r.Put("/{id}", orgH.Update)
			r.Delete("/{id}", orgH.Delete)
			r.Post("/{id}/join", orgH.Join)
			r.Post("/{id}/leave", orgH.Leave)
			r.Post("/{id}/members", orgH.AddMember)
			r.Delete("/{id}/members/{userId}", orgH.RemoveMember)
			r.Get("/my", orgH.MyOrganizations)
		})
	})

	r.Route("/api/organizations/{orgId}/classes", func(r chi.Router) {
		r.Get("/", classH.List)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", classH.Create)
		})
	})

	r.Route("/api/classes", func(r chi.Router) {
		r.Get("/{id}", classH.GetByID)
		r.Get("/{id}/members", classH.GetMembers)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Put("/{id}", classH.Update)
			r.Delete("/{id}", classH.Delete)
			r.Post("/join", classH.JoinByCode)
			r.Post("/{id}/leave", classH.Leave)
		})
	})

	r.Route("/api/training", func(r chi.Router) {
		r.Get("/", trainingH.List)
		r.Get("/{id}", trainingH.GetByID)
		r.Get("/{id}/enrollments", trainingH.GetEnrollments)
		r.Group(func(r chi.Router) {
			r.Use(middleware.AuthMiddleware(jwtManager))
			r.Post("/", trainingH.Create)
			r.Put("/{id}", trainingH.Update)
			r.Delete("/{id}", trainingH.Delete)
			r.Post("/{id}/enroll", trainingH.Enroll)
			r.Delete("/{id}/enroll", trainingH.Unenroll)
			r.Get("/{id}/progress", trainingH.GetMyProgress)
			r.Post("/{id}/sections", trainingH.AddSection)
			r.Delete("/sections/{sectionId}", trainingH.DeleteSection)
			r.Post("/sections/{sectionId}/problems", trainingH.AddProblem)
			r.Delete("/problems/{problemId}", trainingH.RemoveProblem)
		})
	})

	r.Route("/api/contests/{contestId}/plagiarism", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Post("/check", plagiarismH.RunCheck)
		r.Get("/report", plagiarismH.GetReportByContest)
		r.Get("/report/{reportId}", plagiarismH.GetReport)
		r.Get("/report/{reportId}/pairs", plagiarismH.ListPairs)
		r.Put("/pairs/{pairId}", plagiarismH.UpdatePairStatus)
	})

	r.Route("/api/generate", func(r chi.Router) {
		r.Use(middleware.AuthMiddleware(jwtManager))
		r.Use(middleware.RequireRole("admin", "setter"))
		r.Post("/problem", d.Generate.Problem)
	})

	fileServer := http.FileServer(http.Dir("./media"))
	r.Handle("/media/*", http.StripPrefix("/media/", fileServer))

	return r
}